pumpMng is now a DecimalField instead of IntegerField. Allows for half numbers, eg. 2.5

+++ backend/booking/models.py
@@ -33,7 +33,7 @@ class Bokning(models.Model):
     klient = models.ForeignKey('Klient', on_delete=models.CASCADE)
     slangStr = models.IntegerField(null=True, blank=True)
     maskinist = models.ForeignKey(User, on_delete=models.CASCADE)
-    pumpMng = models.IntegerField()
+    pumpMng = models.DecimalField()
     datum = models.DateField()
     littNr = models.CharField(max_length=124)
     resTid = models.IntegerField()
